server-args-parser(move-module,mechanical_provable): move config parser under utils
