+++ test/manual/test_config_integration.py
@@ -11,7 +11,7 @@ import pytest
 import yaml
 
 from sglang.srt.server_args import ServerArgs, prepare_server_args
-from sglang.srt.server_args_config_parser import ConfigArgumentMerger
+from sglang.srt.utils.server_args_config_parser import ConfigArgumentMerger
 
 
 @pytest.fixture
